[FEAT] #13 - datasource 제거 및 Repository가 datasource 역할 담당
