+++ core/data/build.gradle.kts
@@ -1,5 +1,6 @@
 plugins {
-    id ("haroo.android.library")
+    id("haroo.android.library")
+    id("haroo.android.hilt")
 }
 
 android {
@@ -7,5 +8,8 @@ android {
 }
 
 dependencies {
+    implementation(project(":core:datastore"))
+    implementation(project(":core:database"))
+
     implementation(libs.paging.runtime)
 }
