+++ core/data/src/main/java/com/core/data/image/datastore/ImageDataStoreRepositoryImpl.kt
@@ -1,26 +1,26 @@
-package com.core.datasource.local.image
+package com.core.data.image.datastore
 
 import androidx.paging.Pager
 import androidx.paging.PagingConfig
 import androidx.paging.PagingData
-import com.core.datasource.ImageDatasource
-import com.core.datasource.model.Image
+import com.core.data.image.ImageRepository
+import com.core.data.model.Image
 import com.core.datastore.ImageDatastore
 import kotlinx.coroutines.flow.Flow
 import javax.inject.Inject
 
-class ImageLocalDataSourceImpl @Inject constructor(
+class ImageDataStoreRepositoryImpl @Inject constructor(
     private val imageDatastore: ImageDatastore
-) : ImageDatasource {
+) : ImageRepository {
     override fun getImages(): Flow<PagingData<Image>> {
         return Pager(
             config = PagingConfig(
-                pageSize = ImagePagingSource.PAGING_SIZE,
+                pageSize = ImageDataStorePagingSource.PAGING_SIZE,
                 enablePlaceholders = true,
-                maxSize = ImagePagingSource.PAGING_SIZE * 5
+                maxSize = ImageDataStorePagingSource.PAGING_SIZE * 5
             ),
             pagingSourceFactory = {
-                ImagePagingSource(imageDatastore)
+                ImageDataStorePagingSource(imageDatastore)
             }
         ).flow
     }
