+++ core/data/src/main/java/com/core/data/model/Tag.kt
@@ -1,4 +1,4 @@
-package com.core.datasource.model
+package com.core.data.model
 
 import com.core.database.model.TagEntity
