--- core/datasource/src/main/java/com/core/datasource/di/DatasourceModule.kt
@@ -1,24 +1,0 @@
-package com.core.datasource.di
-
-import com.core.datasource.ImageDatasource
-import com.core.datasource.PostDatasource
-import com.core.datasource.local.image.ImageLocalDataSourceImpl
-import com.core.datasource.local.post.PostLocalDataSourceImpl
-import dagger.Binds
-import dagger.Module
-import dagger.hilt.InstallIn
-import dagger.hilt.components.SingletonComponent
-
-@Module
-@InstallIn(SingletonComponent::class)
-interface DatasourceModule {
-    @Binds
-    fun bindPostDatasource(
-        postLocalDataSourceImpl: PostLocalDataSourceImpl
-    ): PostDatasource
-
-    @Binds
-    fun bindImageDatasource(
-        imageLocalDataSourceImpl: ImageLocalDataSourceImpl
-    ): ImageDatasource
-}
