--- core/datasource/src/test/java/com/core/datasource/ExampleUnitTest.kt
@@ -1,4 +1,0 @@
-package com.core.datasource
-
-class ExampleUnitTest {
-}
