+++ settings.gradle
@@ -19,7 +19,6 @@ include ':core:database'
 include ':core:data'
 include ':core:domain'
 include ':core:datastore'
-include ':core:datasource'
 include ':core:designsystem'
 include ':core:ui'
 include ':core:model'
